# apps/api/src/server.ts
import { createServer, type IncomingMessage, type ServerResponse } from "node:http";
import { randomUUID } from "node:crypto";
import { RelayError } from "../../../packages/core/src/errors.ts";
import { RelayService } from "../../../packages/core/src/relay-service.ts";
import { parseChatCompletionRequest, parseProviderValidationRequest, parseRouteQuery } from "../../../packages/core/src/validation.ts";
import { defaultProviderConfig, defaultRoute, FixedClock, InMemoryRelayStore, InMemoryUsageRepository, OpenAiCompatibleHttpAdapter, SequentialIdGenerator, StaticSecretResolver } from "../../../packages/adapters/src/index.ts";
import { authenticate, validateRuntimeAuthMode } from "./auth.ts";

export function buildDefaultService(): RelayService {
  const provider = defaultProviderConfig();
  const store = new InMemoryRelayStore([defaultRoute()], [provider]);
  return new RelayService({
    routes: store,
    secrets: new StaticSecretResolver(new Map([[provider.secretReference, process.env.RELAY_PROVIDER_API_KEY ?? "dev-placeholder"]])),
    provider: new OpenAiCompatibleHttpAdapter(),
    audit: store,
    usage: new InMemoryUsageRepository(store),
    idempotency: store,
    clock: new FixedClock(new Date()),
    ids: new SequentialIdGenerator(),
  });
}

export function createRelayHttpServer(service: RelayService = buildDefaultService()) {
  validateRuntimeAuthMode();
  return createServer(async (req, res) => {
    try {
      await handleRequest(req, res, service);
    } catch (error) {
      writeError(res, error, req.headers["x-correlation-id"]);
    }
  });
}

async function handleRequest(req: IncomingMessage, res: ServerResponse, service: RelayService): Promise<void> {
  const url = new URL(req.url ?? "/", "http://127.0.0.1");
  if (req.method === "GET" && url.pathname === "/health") {
    writeJson(res, 200, { data: { status: "ok" }, meta: meta(req) });
    return;
  }
  if (req.method === "GET" && url.pathname === "/ready") {
    writeJson(res, 200, { data: { status: "ready" }, meta: meta(req) });
    return;
  }

  const auth = authenticate(singleHeader(req.headers.authorization), singleHeader(req.headers["x-tenant-id"]));
  const ctx = {
    auth,
    requestId: randomUUID(),
    correlationId: singleHeader(req.headers["x-correlation-id"]) ?? randomUUID(),
    now: new Date(),
  };

  if (req.method === "GET" && url.pathname === "/v1/routes/resolve") {
    const query = parseRouteQuery(url.searchParams);
    const resolution = await service.resolve(ctx, query);
    writeJson(res, resolution.allowed ? 200 : 403, { data: redactedResolution(resolution), meta: meta(req, ctx.correlationId) });
    return;
  }

  if (req.method === "POST" && url.pathname === "/v1/chat/completions") {
    const body = parseChatCompletionRequest(await readJson(req));
    const idempotencyKey = singleHeader(req.headers["idempotency-key"]) ?? "";
    const response = await service.completeChat(ctx, body, idempotencyKey);
    writeJson(res, 200, { data: response, meta: meta(req, ctx.correlationId) });
    return;
  }

  if (req.method === "GET" && url.pathname === "/v1/usage") {
    const records = await service.listUsage(ctx);
    writeJson(res, 200, { data: { usage: records.map(redactedUsage) }, meta: meta(req, ctx.correlationId) });
    return;
  }

  if (req.method === "POST" && url.pathname === "/v1/providers/validate") {
    const body = parseProviderValidationRequest(await readJson(req));
    const result = await service.validateProvider(ctx, body);
    writeJson(res, result.valid ? 200 : 422, { data: result, meta: meta(req, ctx.correlationId) });
    return;
  }

  throw new RelayError("VALIDATION_FAILED", "Route not found.", 404);
}

function redactedUsage(record: Awaited<ReturnType<RelayService["listUsage"]>>[number]): Record<string, unknown> {
  return {
    id: record.id,
    requestId: record.requestId,
    routeId: record.routeId,
    providerId: record.providerId,
    model: record.model,
    inputTokens: record.usage.inputTokens,
    outputTokens: record.usage.outputTokens,
    estimatedCostCents: record.usage.estimatedCostCents,
    latencyMs: record.latencyMs,
    terminalReason: record.terminalReason,
    correlationId: record.correlationId,
    createdAt: record.createdAt.toISOString(),
  };
}

function redactedResolution(resolution: Awaited<ReturnType<RelayService["resolve"]>>): Record<string, unknown> {
  return {
    allowed: resolution.allowed,
    reasonCodes: resolution.reasonCodes,
    route: resolution.route === undefined ? undefined : {
      routeId: resolution.route.routeId,
      providerId: resolution.route.providerId,
      model: resolution.route.model,
      requiredCapabilities: resolution.route.requiredCapabilities,
      maxCostCents: resolution.route.maxCostCents,
    },
    provider: resolution.provider === undefined ? undefined : {
      providerId: resolution.provider.providerId,
      adapterType: resolution.provider.adapterType,
      capabilities: resolution.provider.capabilities,
    },
  };
}

async function readJson(req: IncomingMessage): Promise<unknown> {
  const chunks: Buffer[] = [];
  for await (const chunk of req) {
    chunks.push(Buffer.isBuffer(chunk) ? chunk : Buffer.from(chunk));
  }
  const raw = Buffer.concat(chunks).toString("utf8");
  return raw.length === 0 ? {} : JSON.parse(raw);
}

function writeJson(res: ServerResponse, status: number, body: unknown): void {
  res.writeHead(status, { "content-type": "application/json" });
  res.end(JSON.stringify(body));
}

function writeError(res: ServerResponse, error: unknown, correlationHeader: string | string[] | undefined): void {
  const correlationId = singleHeader(correlationHeader) ?? randomUUID();
  if (error instanceof RelayError) {
    writeJson(res, error.status, {
      error: {
        code: error.code,
        message: error.message,
        details: error.details,
        correlationId,
      },
    });
    return;
  }
  writeJson(res, 500, {
    error: {
      code: "DEPENDENCY_UNAVAILABLE",
      message: "Unexpected failure.",
      details: [],
      correlationId,
    },
  });
}

function meta(req: IncomingMessage, correlationId = singleHeader(req.headers["x-correlation-id"]) ?? randomUUID()): Record<string, string> {
  return {
    requestId: randomUUID(),
    correlationId,
    apiVersion: "v1",
  };
}

function singleHeader(value: string | string[] | undefined): string | undefined {
  return Array.isArray(value) ? value[0] : value;
}